requirements.txt
feedparser==6.0.8
newspaper3k==0.2.8
bert_extractive_summarizer==0.8.1
Jinja2==3.0.3
newspaper==0.1.0.7
PyGithub==1.55
summarizer==0.0.7
